\name{NHMpkg-package}
\alias{NHMpkg-package}
\alias{NHMpkg}
\docType{package}
\title{
\packageTitle{NHMpkg}
}
\description{
\packageDescription{NHMpkg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NHMpkg}
\packageIndices{NHMpkg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{NHMpkg}

Maintainer: \packageMaintainer{NHMpkg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
